Adds response wrapper struct for Redditor.
This adds some godoc for Agent and adds a file for wrapper structs which embed
the expected protobuffer return and accomodate error strings.

There might be a better way to do that I discover in the future, but I see no
sense in having an error string in the protobuffers that is not relevant to the
data itself.