Reorganizes nface for coherence and testing.
The need for a client and user agent to persist between requests invalidated the
old design of giving requests behaviors. nface.Request is now a passive data
structure employed by nface.Client, which remembers the persistent attributes.

response.go is no longer necessary because nface now checks for http status; an
error field will not be returned under a status of 200/OK, and the error will be
caught upstream from Agent's api functions.