Increase rate limit.
Setting the rate limit to 60 queries per second causes some problems in my
tests. The time is probably not exact and 60qps is being exceeded.

For non-OAuth2 clients the limit is 30qps, which seems life a safe point for
now. A future commit should read from Reddit's responses to determine the number
of queries remaining for the time window.

This will halve the speed of graw queries, but that's better than being
unreliable. I intend to revise the qeury batching before v1 anyway.