Implement a scanner for listings in general.
This moves the processing of a subreddit's link listing endpoint which was done
in postmonitor to a new package called scanner. This will scan listing endpoints
of any kind--not necessarily link listings.

A future commit will implement a user monitor, and take advantage of this
generalized scanner to monitor a user's comments.


Former-commit-id: 2d7661ce863967010a8ed225f04d218bb1e4e93c