[No build] Updates scanner for operator changes.
This commit updates scanner for the changes to operator. While updating this, I
realized that some things which depend on scanner were actually modifying
scanner's behavior via a mock operator, which scanner depends on. That's crazy;
so I did the same thing for scanner that I did for operator: provide a mock
implementation. I'll resolve that nonsense next.


Former-commit-id: eccbec26aff254c586da14596d19ca869f402c30