[No build] Update operator for redditproto 2.
This updates operator to take advantage of the improved api to redditproto's
listing parsing utilities. It is a partial implementation and doesn't build
outside the operator package.

This is the first of a series of no-builds that implement a backwards crawling
engine. This commit enableds monitors to more easily share default behavior,
which the next commit will implement.