[No build] Move Handlers to monitor package.
There is no reason for these be part of the public api. If necessary I can
export a check function or similar later, but exporting for the sake of users
testing by interface cast is a lackluster offering and it gets in my way.

The receivers for some of the inbox functions are changed to receive comments
because Reddit only pretends they are messages, and the new redditproto parsers
isn't as gullible.


Former-commit-id: 79740346511bc40c1298fd55e8296cfa9c98649d