[No build] Unexports api package.
I've been thinking over whether to continue exporting the api package for a
while, and have finally decided against it. The only necessary thing for bots
from the package was the Engine api, which can be exported from graw itself as
it is in this commit.

Two packages are already required for import to use graw (graw and redditproto);
I think a third package that isn't entirely necessary is an annoyance more than
anything.

The documentation for the api package will still be available on godoc even
though it is internal.