Manage engines from graw top-level package.
This commit provides the engine interface that bots want from the graw package
so no other packages are necessary to import for it (what a previous commit
failed to do).

graw now has runner methods for both forward time and backwards time engines.

graw now provides a method for accessing the engine running a given bot that
doesn't require the bot implement a method to do it. Bots can claim it in SetUp
if needed.


Former-commit-id: 7a2112843a0b242e36fb88c61046cacdd59a658c