Fix event generation for private messages.
The /message/messages endpoint was not similar to the others (e.g. selfreply).
It would provide messages in a listing, but each listing child was a full
sequence of messages rather than an individual message in order or time
received. It even included sent messages.

The api call for checking whether things exist doesn't work for messages. Those
have to be checked by id only from the /message/messages/<id> endpoint, which
operator now does in IsThereThing.