Implement skeleton for end to end testing.
For the moment this is mostly just for me; I don't plan to outline how to set up
the end to end testing environment. But in https://github.com/turnage/graw-test
I plan to automate the setup so anyone can just download the vagrant for it.
That's a work in progress.


Former-commit-id: 5bd3c8bc9a79d06e308de26f0b0de5c6c9cb993c