Stop reddit from sending HTML entities in JSON
In responses to API calls, reddit apparently converts a few characters
to HTML entites. The API documentation states:

https://www.reddit.com/dev/api#response_body_encoding
> For legacy reasons, all JSON response bodies currently have <, >, and
> & replaced with &lt;, &gt;, and &amp;, respectively. If you wish to
> opt out of this behaviour, add a raw_json=1 parameter to your request.

Such behavior offers no benefit to users of graw, especially so
considering that graw is (as far as I understand) intended to be a nice
abstraction that hides gotchas like this. So this commit tells graw to
pass raw_json=1 in every GET request.

For graw bots that expect <, >, and & delivered as HTML entites, this is
a breaking change.


Former-commit-id: 548cc9ca47477d0f5b587a404c4f1d93802f1e2e